backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
APScheduler==3.10.4
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
tensorflow==2.17.1
firebase-admin==6.5.0
Werkzeug==3.0.2
